--- a/src/concurrency/data_race.rs
+++ b/src/concurrency/data_race.rs
@@ -50,8 +50,11 @@ use rustc_ast::Mutability;
 use rustc_data_structures::fx::{FxHashMap, FxHashSet};
 use rustc_index::vec::{Idx, IndexVec};
 use rustc_middle::mir;
+use rustc_span::Span;
+use rustc_span::DUMMY_SP;
 use rustc_target::abi::{Align, Size};
 
+use crate::diagnostics::RacingOp;
 use crate::*;
 
 use super::{
@@ -144,8 +147,8 @@ impl ThreadClockSet {
     /// Increment the happens-before clock at a
     /// known index.
     #[inline]
-    fn increment_clock(&mut self, index: VectorIdx) {
-        self.clock.increment_index(index);
+    fn increment_clock(&mut self, index: VectorIdx, current_span: Span) {
+        self.clock.increment_index(index, current_span);
     }
 
     /// Join the happens-before clock with that of
@@ -361,6 +364,8 @@ impl MemoryCellClocks {
             atomic.read_vector.set_at_index(&clocks.clock, index);
             Ok(())
         } else {
+            let atomic = self.atomic_mut();
+            atomic.read_vector.set_at_index(&clocks.clock, index);
             Err(DataRace)
         }
     }
@@ -378,6 +383,8 @@ impl MemoryCellClocks {
             atomic.write_vector.set_at_index(&clocks.clock, index);
             Ok(())
         } else {
+            let atomic = self.atomic_mut();
+            atomic.write_vector.set_at_index(&clocks.clock, index);
             Err(DataRace)
         }
     }
@@ -386,46 +393,51 @@ impl MemoryCellClocks {
     /// returns true if a data-race is detected.
     fn read_race_detect(
         &mut self,
-        clocks: &ThreadClockSet,
+        clocks: &mut ThreadClockSet,
         index: VectorIdx,
+        current_span: Span,
     ) -> Result<(), DataRace> {
         log::trace!("Unsynchronized read with vectors: {:#?} :: {:#?}", self, clocks);
-        if self.write <= clocks.clock[self.write_index] {
+        let res = if self.write <= clocks.clock[self.write_index] {
             let race_free = if let Some(atomic) = self.atomic() {
                 atomic.write_vector <= clocks.clock
             } else {
                 true
             };
-            if race_free {
-                self.read.set_at_index(&clocks.clock, index);
-                Ok(())
-            } else {
-                Err(DataRace)
-            }
+            self.read.set_at_index(&clocks.clock, index);
+            if race_free { Ok(()) } else { Err(DataRace) }
         } else {
             Err(DataRace)
+        };
+        if res.is_ok() && current_span != DUMMY_SP {
+            clocks.clock[index].span = current_span;
         }
+        res
     }
 
     /// Detect races for non-atomic write operations at the current memory cell
     /// returns true if a data-race is detected.
     fn write_race_detect(
         &mut self,
-        clocks: &ThreadClockSet,
+        clocks: &mut ThreadClockSet,
         index: VectorIdx,
         write_type: WriteType,
+        current_span: Span,
     ) -> Result<(), DataRace> {
         log::trace!("Unsynchronized write with vectors: {:#?} :: {:#?}", self, clocks);
+        if current_span != DUMMY_SP {
+            clocks.clock[index].span = current_span;
+        }
         if self.write <= clocks.clock[self.write_index] && self.read <= clocks.clock {
             let race_free = if let Some(atomic) = self.atomic() {
                 atomic.write_vector <= clocks.clock && atomic.read_vector <= clocks.clock
             } else {
                 true
             };
+            self.write = clocks.clock[index];
+            self.write_index = index;
+            self.write_type = write_type;
             if race_free {
-                self.write = clocks.clock[index];
-                self.write_index = index;
-                self.write_type = write_type;
                 self.read.set_zero_vector();
                 Ok(())
             } else {
@@ -621,30 +633,35 @@ pub trait EvalContextExt<'mir, 'tcx: 'mir>: MiriInterpCxExt<'mir, 'tcx> {
     /// Update the data-race detector for an atomic fence on the current thread.
     fn atomic_fence(&mut self, atomic: AtomicFenceOrd) -> InterpResult<'tcx> {
         let this = self.eval_context_mut();
+        let current_span = this.machine.current_span();
         if let Some(data_race) = &mut this.machine.data_race {
-            data_race.maybe_perform_sync_operation(&this.machine.threads, |index, mut clocks| {
-                log::trace!("Atomic fence on {:?} with ordering {:?}", index, atomic);
+            data_race.maybe_perform_sync_operation(
+                &this.machine.threads,
+                |index, mut clocks| {
+                    log::trace!("Atomic fence on {:?} with ordering {:?}", index, atomic);
 
-                // Apply data-race detection for the current fences
-                // this treats AcqRel and SeqCst as the same as an acquire
-                // and release fence applied in the same timestamp.
-                if atomic != AtomicFenceOrd::Release {
-                    // Either Acquire | AcqRel | SeqCst
-                    clocks.apply_acquire_fence();
-                }
-                if atomic != AtomicFenceOrd::Acquire {
-                    // Either Release | AcqRel | SeqCst
-                    clocks.apply_release_fence();
-                }
-                if atomic == AtomicFenceOrd::SeqCst {
-                    data_race.last_sc_fence.borrow_mut().set_at_index(&clocks.clock, index);
-                    clocks.fence_seqcst.join(&data_race.last_sc_fence.borrow());
-                    clocks.write_seqcst.join(&data_race.last_sc_write.borrow());
-                }
+                    // Apply data-race detection for the current fences
+                    // this treats AcqRel and SeqCst as the same as an acquire
+                    // and release fence applied in the same timestamp.
+                    if atomic != AtomicFenceOrd::Release {
+                        // Either Acquire | AcqRel | SeqCst
+                        clocks.apply_acquire_fence();
+                    }
+                    if atomic != AtomicFenceOrd::Acquire {
+                        // Either Release | AcqRel | SeqCst
+                        clocks.apply_release_fence();
+                    }
+                    if atomic == AtomicFenceOrd::SeqCst {
+                        data_race.last_sc_fence.borrow_mut().set_at_index(&clocks.clock, index);
+                        clocks.fence_seqcst.join(&data_race.last_sc_fence.borrow());
+                        clocks.write_seqcst.join(&data_race.last_sc_write.borrow());
+                    }
 
-                // Increment timestamp in case of release semantics.
-                Ok(atomic != AtomicFenceOrd::Acquire)
-            })
+                    // Increment timestamp in case of release semantics.
+                    Ok(atomic != AtomicFenceOrd::Acquire)
+                },
+                current_span,
+            )
         } else {
             Ok(())
         }
@@ -682,6 +699,7 @@ impl VClockAlloc {
         thread_mgr: &ThreadManager<'_, '_>,
         len: Size,
         kind: MemoryKind<MiriMemoryKind>,
+        current_span: Span,
     ) -> VClockAlloc {
         let (alloc_timestamp, alloc_index) = match kind {
             // User allocated and stack memory should track allocation.
@@ -693,7 +711,10 @@ impl VClockAlloc {
             )
             | MemoryKind::Stack => {
                 let (alloc_index, clocks) = global.current_thread_state(thread_mgr);
-                let alloc_timestamp = clocks.clock[alloc_index];
+                let mut alloc_timestamp = clocks.clock[alloc_index];
+                if current_span != DUMMY_SP {
+                    alloc_timestamp.span = current_span;
+                }
                 (alloc_timestamp, alloc_index)
             }
             // Other global memory should trace races but be allocated at the 0 timestamp.
@@ -704,7 +725,7 @@ impl VClockAlloc {
                 | MiriMemoryKind::ExternStatic
                 | MiriMemoryKind::Tls,
             )
-            | MemoryKind::CallerLocation => (0, VectorIdx::MAX_INDEX),
+            | MemoryKind::CallerLocation => (VTimestamp::NONE, VectorIdx::MAX_INDEX),
         };
         VClockAlloc {
             alloc_ranges: RefCell::new(RangeMap::new(
@@ -735,7 +756,7 @@ impl VClockAlloc {
                     let idx = l_remainder_slice
                         .iter()
                         .enumerate()
-                        .find_map(|(idx, &r)| if r == 0 { None } else { Some(idx) })
+                        .find_map(|(idx, &r)| if r == VTimestamp::NONE { None } else { Some(idx) })
                         .expect("Invalid VClock Invariant");
                     Some(idx + r_slice.len())
                 } else {
@@ -762,7 +783,7 @@ impl VClockAlloc {
     ) -> InterpResult<'tcx> {
         let (current_index, current_clocks) = global.current_thread_state(thread_mgr);
         let write_clock;
-        let (other_action, other_thread, _other_clock) = if range.write
+        let (other_action, other_thread, other_clock) = if range.write
             > current_clocks.clock[range.write_index]
         {
             // Convert the write action into the vector clock it
@@ -799,14 +820,19 @@ impl VClockAlloc {
         let other_thread_info = global.print_thread_metadata(thread_mgr, other_thread);
 
         // Throw the data-race detection.
-        throw_ub_format!(
-            "Data race detected between {} on {} and {} on {} at {:?}",
-            action,
-            current_thread_info,
-            other_action,
-            other_thread_info,
-            ptr_dbg,
-        )
+        Err(err_machine_stop!(TerminationInfo::DataRace {
+            ptr: ptr_dbg,
+            op1: RacingOp {
+                action: action.to_string(),
+                thread_info: current_thread_info,
+                span: current_clocks.clock.as_slice()[current_index.index()].span_data(),
+            },
+            op2: RacingOp {
+                action: other_action.to_string(),
+                thread_info: other_thread_info,
+                span: other_clock.as_slice()[other_thread.index()].span_data(),
+            },
+        }))?
     }
 
     /// Detect racing atomic read and writes (not data races)
@@ -840,12 +866,14 @@ impl VClockAlloc {
         range: AllocRange,
         machine: &MiriMachine<'_, '_>,
     ) -> InterpResult<'tcx> {
+        let current_span = machine.current_span();
         let global = machine.data_race.as_ref().unwrap();
         if global.race_detecting() {
-            let (index, clocks) = global.current_thread_state(&machine.threads);
+            let (index, mut clocks) = global.current_thread_state_mut(&machine.threads);
             let mut alloc_ranges = self.alloc_ranges.borrow_mut();
             for (offset, range) in alloc_ranges.iter_mut(range.start, range.size) {
-                if let Err(DataRace) = range.read_race_detect(&clocks, index) {
+                if let Err(DataRace) = range.read_race_detect(&mut clocks, index, current_span) {
+                    drop(clocks);
                     // Report data-race.
                     return Self::report_data_race(
                         global,
@@ -871,11 +899,15 @@ impl VClockAlloc {
         write_type: WriteType,
         machine: &mut MiriMachine<'_, '_>,
     ) -> InterpResult<'tcx> {
+        let current_span = machine.current_span();
         let global = machine.data_race.as_mut().unwrap();
         if global.race_detecting() {
-            let (index, clocks) = global.current_thread_state(&machine.threads);
+            let (index, mut clocks) = global.current_thread_state_mut(&machine.threads);
             for (offset, range) in self.alloc_ranges.get_mut().iter_mut(range.start, range.size) {
-                if let Err(DataRace) = range.write_race_detect(&clocks, index, write_type) {
+                if let Err(DataRace) =
+                    range.write_race_detect(&mut clocks, index, write_type, current_span)
+                {
+                    drop(clocks);
                     // Report data-race
                     return Self::report_data_race(
                         global,
@@ -1100,6 +1132,7 @@ trait EvalContextPrivExt<'mir, 'tcx: 'mir>: MiriInterpCxExt<'mir, 'tcx> {
                     size.bytes()
                 );
 
+                let current_span = this.machine.current_span();
                 // Perform the atomic operation.
                 data_race.maybe_perform_sync_operation(
                     &this.machine.threads,
@@ -1124,6 +1157,7 @@ trait EvalContextPrivExt<'mir, 'tcx: 'mir>: MiriInterpCxExt<'mir, 'tcx> {
                         // This conservatively assumes all operations have release semantics
                         Ok(true)
                     },
+                    current_span,
                 )?;
 
                 // Log changes to atomic memory.
@@ -1303,7 +1337,12 @@ impl GlobalState {
     // Hook for thread creation, enabled multi-threaded execution and marks
     // the current thread timestamp as happening-before the current thread.
     #[inline]
-    pub fn thread_created(&mut self, thread_mgr: &ThreadManager<'_, '_>, thread: ThreadId) {
+    pub fn thread_created(
+        &mut self,
+        thread_mgr: &ThreadManager<'_, '_>,
+        thread: ThreadId,
+        current_span: Span,
+    ) {
         let current_index = self.current_index(thread_mgr);
 
         // Enable multi-threaded execution, there are now at least two threads
@@ -1320,7 +1359,7 @@ impl GlobalState {
             // Now re-configure the re-use candidate, increment the clock
             // for the new sync use of the vector.
             let vector_clocks = self.vector_clocks.get_mut();
-            vector_clocks[reuse_index].increment_clock(reuse_index);
+            vector_clocks[reuse_index].increment_clock(reuse_index, current_span);
 
             // Locate the old thread the vector was associated with and update
             // it to represent the new thread instead.
@@ -1360,8 +1399,8 @@ impl GlobalState {
 
         // Advance both threads after the synchronized operation.
         // Both operations are considered to have release semantics.
-        current.increment_clock(current_index);
-        created.increment_clock(created_index);
+        current.increment_clock(current_index, current_span);
+        created.increment_clock(created_index, current_span);
     }
 
     /// Hook on a thread join to update the implicit happens-before relation between the joined
@@ -1427,13 +1466,13 @@ impl GlobalState {
     /// This should be called strictly before any calls to
     /// `thread_joined`.
     #[inline]
-    pub fn thread_terminated(&mut self, thread_mgr: &ThreadManager<'_, '_>) {
+    pub fn thread_terminated(&mut self, thread_mgr: &ThreadManager<'_, '_>, current_span: Span) {
         let current_index = self.current_index(thread_mgr);
 
         // Increment the clock to a unique termination timestamp.
         let vector_clocks = self.vector_clocks.get_mut();
         let current_clocks = &mut vector_clocks[current_index];
-        current_clocks.increment_clock(current_index);
+        current_clocks.increment_clock(current_index, current_span);
 
         // Load the current thread id for the executing vector.
         let vector_info = self.vector_info.get_mut();
@@ -1463,12 +1502,13 @@ impl GlobalState {
         &self,
         thread_mgr: &ThreadManager<'_, '_>,
         op: impl FnOnce(VectorIdx, RefMut<'_, ThreadClockSet>) -> InterpResult<'tcx, bool>,
+        current_span: Span,
     ) -> InterpResult<'tcx> {
         if self.multi_threaded.get() {
             let (index, clocks) = self.current_thread_state_mut(thread_mgr);
             if op(index, clocks)? {
                 let (_, mut clocks) = self.current_thread_state_mut(thread_mgr);
-                clocks.increment_clock(index);
+                clocks.increment_clock(index, current_span);
             }
         }
         Ok(())
@@ -1501,10 +1541,10 @@ impl GlobalState {
     /// since an acquire operation should have occurred before, however
     /// for futex & condvar operations this is not the case and this
     /// operation must be used.
-    pub fn validate_lock_release(&self, lock: &mut VClock, thread: ThreadId) {
+    pub fn validate_lock_release(&self, lock: &mut VClock, thread: ThreadId, current_span: Span) {
         let (index, mut clocks) = self.load_thread_state_mut(thread);
         lock.clone_from(&clocks.clock);
-        clocks.increment_clock(index);
+        clocks.increment_clock(index, current_span);
     }
 
     /// Release a lock handle, express that this happens-before
@@ -1514,10 +1554,15 @@ impl GlobalState {
     /// For normal locks this should be equivalent to `validate_lock_release`.
     /// This function only exists for joining over the set of concurrent readers
     /// in a read-write lock and should not be used for anything else.
-    pub fn validate_lock_release_shared(&self, lock: &mut VClock, thread: ThreadId) {
+    pub fn validate_lock_release_shared(
+        &self,
+        lock: &mut VClock,
+        thread: ThreadId,
+        current_span: Span,
+    ) {
         let (index, mut clocks) = self.load_thread_state_mut(thread);
         lock.join(&clocks.clock);
-        clocks.increment_clock(index);
+        clocks.increment_clock(index, current_span);
     }
 
     /// Load the vector index used by the given thread as well as the set of vector clocks