@RalfJung 2022-11-04
I assume this is intended to fix https://github.com/rust-lang/miri/issues/2205 ?