@RalfJung 2022-12-21
@saethlin could you do benchmarks with `-Zmiri-disable-stacked-borrows -Zmiri-disable-validation` so that we can an idea for how expensive this is?