@saethlin 2022-12-22
Before:
```
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/backtraces/Cargo.toml
  Time (mean ± σ):      2.262 s ±  0.005 s    [User: 2.204 s, System: 0.054 s]
  Range (min … max):    2.256 s …  2.270 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/mse/Cargo.toml
  Time (mean ± σ):     389.0 ms ±   4.6 ms    [User: 351.8 ms, System: 36.8 ms]
  Range (min … max):   383.3 ms … 394.1 ms    7 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/serde1/Cargo.toml
  Time (mean ± σ):     930.7 ms ±   6.6 ms    [User: 882.4 ms, System: 47.3 ms]
  Range (min … max):   924.3 ms … 940.7 ms    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/serde2/Cargo.toml
  Time (mean ± σ):      1.844 s ±  0.033 s    [User: 1.795 s, System: 0.047 s]
  Range (min … max):    1.812 s …  1.885 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/slice-get-unchecked/Cargo.toml
  Time (mean ± σ):     322.3 ms ±   4.2 ms    [User: 282.1 ms, System: 40.0 ms]
  Range (min … max):   313.0 ms … 326.0 ms    9 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/unicode/Cargo.toml
  Time (mean ± σ):     717.5 ms ±   2.1 ms    [User: 681.5 ms, System: 35.4 ms]
  Range (min … max):   714.5 ms … 720.2 ms    5 runs
```

After:
```
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/backtraces/Cargo.toml
  Time (mean ± σ):      2.275 s ±  0.038 s    [User: 2.201 s, System: 0.057 s]
  Range (min … max):    2.245 s …  2.341 s    5 runs
  
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/mse/Cargo.toml
  Time (mean ± σ):     393.9 ms ±   1.2 ms    [User: 360.0 ms, System: 33.6 ms]
  Range (min … max):   392.4 ms … 396.1 ms    7 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/serde1/Cargo.toml
  Time (mean ± σ):     933.1 ms ±   3.2 ms    [User: 888.6 ms, System: 43.6 ms]
  Range (min … max):   927.8 ms … 935.5 ms    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/serde2/Cargo.toml
  Time (mean ± σ):      1.909 s ±  0.006 s    [User: 1.851 s, System: 0.056 s]
  Range (min … max):    1.899 s …  1.916 s    5 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/slice-get-unchecked/Cargo.toml
  Time (mean ± σ):     325.8 ms ±   2.0 ms    [User: 287.1 ms, System: 38.5 ms]
  Range (min … max):   323.5 ms … 329.6 ms    9 runs
 
Benchmark 1: cargo +miri miri run --manifest-path /home/ben/miri/bench-cargo-miri/unicode/Cargo.toml
  Time (mean ± σ):     715.4 ms ±   3.4 ms    [User: 672.1 ms, System: 42.6 ms]
  Range (min … max):   711.0 ms … 720.2 ms    5 runs
```

This is about the size of the effect I expected.